Fix: cast a numeric la colonna base della CTE ricorsiva in v_recipe_costs
Postgres rifiutava la vista: il termine non ricorsivo aveva quantity di tipo
numeric(12,4), quello ricorsivo (dopo la divisione per yield_quantity) numeric
senza precisione fissa. I due rami di una UNION ricorsiva devono avere lo
stesso tipo colonna per colonna.

+++ supabase/migrations/20260802000005_recipe_preparations.sql
@@ -115,7 +115,7 @@ with recursive expansion as (
     ri.recipe_id as root_recipe_id,
     ri.ingredient_id,
     ri.component_recipe_id,
-    ri.quantity as multiplier,
+    ri.quantity::numeric as multiplier,
     ri.waste_percentage,
     ri.is_optional,
     1 as depth
